+++ src/main/java/frc/robot/SubsystemManager/SubsystemManager.java
@@ -6,12 +6,13 @@ import static frc.robot.SubsystemManager.SubsystemManagerConstants.*;
 import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.DriverStation.Alliance;
 import frc.robot.Subsystems.Algaer.Algaer;
+import frc.robot.Subsystems.AprilTagVision.AprilTagVision;
 import frc.robot.Subsystems.AutoAlign.AutoAlign;
 import frc.robot.Subsystems.Coraler.Coraler;
 import frc.robot.Subsystems.Drive.Drive;
 import frc.robot.Subsystems.Elevator.Elevator;
 import frc.robot.Subsystems.LED.LED;
-import frc.robot.Subsystems.Vision.AprilTagVision;
+
 import org.littletonrobotics.junction.Logger;
 import org.team7525.subsystem.Subsystem;
