+++ src/main/java/frc/robot/subsystems/AprilTagVision/AprilTagVisionIOSim.java
@@ -1,12 +1,13 @@
-package frc.robot.Subsystems.Vision;
-
-import static frc.robot.Subsystems.Vision.VisionConstants.*;
+package frc.robot.Subsystems.AprilTagVision;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Pose3d;
 import edu.wpi.first.math.geometry.Rotation2d;
+import frc.robot.Subsystems.AprilTagVision.VisionConstants.CameraResolution;
+
+import static frc.robot.Subsystems.AprilTagVision.VisionConstants.*;
+
 import org.photonvision.PhotonCamera;
-import frc.robot.Subsystems.Vision.VisionConstants.CameraResolution;
 import org.photonvision.simulation.PhotonCameraSim;
 import org.photonvision.simulation.SimCameraProperties;
 import org.photonvision.simulation.VisionSystemSim;
